+++ dessia_common/workflow/core.py
@@ -20,15 +20,14 @@ from dessia_common.templates import workflow_template
 from dessia_common import DessiaObject, is_sequence, JSONSCHEMA_HEADER, jsonschema_from_annotation,\
     deserialize_argument, set_default_value, prettyname, serialize_dict, DisplaySetting
 
-from dessia_common.utils.serialization import dict_to_object, deserialize, serialize_with_pointers, serialize,\
-                                              dereference_jsonpointers
+from dessia_common.utils.serialization import deserialize, serialize_with_pointers, serialize, update_pointers_data
 from dessia_common.utils.types import serialize_typing, deserialize_typing, recursive_type, typematch
 from dessia_common.utils.copy import deepcopy_value
 from dessia_common.utils.docstrings import FAILED_ATTRIBUTE_PARSING, EMPTY_PARSED_ATTRIBUTE
 from dessia_common.utils.diff import choose_hash
 from dessia_common.typings import JsonSerializable, MethodType
 from dessia_common.displays import DisplayObject
-from dessia_common.breakdown import attrmethod_getter
+from dessia_common.breakdown import attrmethod_getter, ExtractionError
 
 
 class Variable(DessiaObject):
@@ -575,9 +574,15 @@ class Workflow(Block):
         """
         Recompute the object from a dict
         """
-        blocks = [DessiaObject.dict_to_object(d) for d in dict_['blocks']]
+        global_dict, pointers_memo = update_pointers_data(global_dict=global_dict, current_dict=dict_,
+                                                          pointers_memo=pointers_memo)
+
+        blocks = [deserialize(serialized_element=d, global_dict=global_dict, pointers_memo=pointers_memo)
+                  for d in dict_["blocks"]]
         if 'nonblock_variables' in dict_:
-            nonblock_variables = [dict_to_object(d) for d in dict_['nonblock_variables']]
+            nonblock_variables = [deserialize(serialized_element=d, global_dict=global_dict,
+                                              pointers_memo=pointers_memo)
+                                  for d in dict_['nonblock_variables']]
         else:
             nonblock_variables = []
 
@@ -1016,9 +1021,6 @@ class Workflow(Block):
 
         state.output_value = state.values[self.outputs[0]]
 
-        name_index = str(len(self.inputs) + 1)
-        if name is None and name_index in input_values:
-            name = input_values[name_index]
         if not name:
             timestamp = start_timestamp.strftime("%m-%d (%H:%M)")
             name = f"{self.name} @ [{timestamp}]"
@@ -1138,10 +1140,10 @@ class Workflow(Block):
         classes = []
 
         script_blocks = ''
-        for ib, block in enumerate(self.blocks):
+        for idxb, block in enumerate(self.blocks):
             block_script, classes_block = block.to_script()
             classes.extend(classes_block)
-            script_blocks += f'block_{ib} = {block_script}\n'
+            script_blocks += f'block_{idxb} = {block_script}\n'
 
         script = ('import dessia_common.workflow as dcw\n'
                   + 'import dessia_common.workflow.blocks as dcw_blocks\n'
@@ -1152,9 +1154,9 @@ class Workflow(Block):
             script += f'import {module}\n'
         script += '\n'
         script += script_blocks
-        script += 'blocks = [{}]\n'.format(', '.join(['block_' + str(i) for i in range(len(self.blocks))]))
+        script += f"blocks = [{', '.join(['block_' + str(i) for i in range(len(self.blocks))])}]\n"
 
-        for ip, pipe in enumerate(self.pipes):
+        for idxp, pipe in enumerate(self.pipes):
             input_index = self.variable_indices(pipe.input_variable)
             if isinstance(input_index, int):
                 script += pipe.input_variable.to_script(variable_index=variable_index) + '\n'
@@ -1170,7 +1172,7 @@ class Workflow(Block):
                 variable_index += 1
             else:
                 output_name = f"block_{output_index[0]}.inputs[{output_index[2]}]" + '\n'
-            script += pipe.to_script(pipe_index=ip, input_name=input_name, output_name=output_name) + '\n'
+            script += pipe.to_script(pipe_index=idxp, input_name=input_name, output_name=output_name) + '\n'
 
         script += f"pipes = [{', '.join(['pipe_' + str(i) for i in range(len(self.pipes))])}]\n"
 
@@ -1424,20 +1426,14 @@ class WorkflowState(DessiaObject):
     def dict_to_object(cls, dict_: JsonSerializable, force_generic: bool = False,
                        global_dict=None, pointers_memo: Dict[str, Any] = None, path: str = '#') -> 'WorkflowState':
 
-        # This is copy pasted from generic dict to object, because it is difficult to do a decorator
-        # for both classmethod and function
-        if pointers_memo is None:
-            pointers_memo = {}
-
-        if global_dict is None:
-            global_dict = dict_
-            pointers_memo.update(dereference_jsonpointers(dict_))
+        global_dict, pointers_memo = update_pointers_data(global_dict=global_dict, current_dict=dict_,
+                                                          pointers_memo=pointers_memo)
 
         workflow = Workflow.dict_to_object(dict_['workflow'])
         if 'output_value' in dict_:  # and 'output_value_type' in dict_:
             # type_ = dict_['output_value_type']
             value = dict_['output_value']
-            output_value = deserialize(value, global_dict=dict_,
+            output_value = deserialize(value, global_dict=global_dict,
                                        pointers_memo=pointers_memo, path=f'{path}/output_value')
         else:
             output_value = None
@@ -1445,23 +1441,22 @@ class WorkflowState(DessiaObject):
         values = {}
         if 'values' in dict_:
             for i, value in dict_['values'].items():
-                values[workflow.variables[int(i)]] = deserialize(value, global_dict=dict_,
-                                                                 pointers_memo=pointers_memo,
-                                                                 path=f'{path}/values/{i}')
+                values[workflow.variables[int(i)]] = deserialize(value, global_dict=global_dict,
+                                                                 pointers_memo=pointers_memo, path=f'{path}/values/{i}')
         # elif 'variable_values' in dict_:
         #     for i, value in dict_['variable_values'].items():
         #         values[workflow.variables[int(i)]] = deserialize(value, global_dict=dict_,
         #                                                          pointers_memo=pointers_memo,
         #                                                          path=f'{path}/variable_values/{i}')
 
-        input_values = {int(i): deserialize(v, global_dict=dict_, pointers_memo=pointers_memo,
+        input_values = {int(i): deserialize(v, global_dict=global_dict, pointers_memo=pointers_memo,
                                             path=f"{path}/input_values/{i}")
                         for i, v in dict_['input_values'].items()}
 
-        activated_items = {b: (True if i in dict_['evaluated_blocks_indices'] else False)
+        activated_items = {b: i in dict_['evaluated_blocks_indices']
                            for i, b in enumerate(workflow.blocks)}
 
-        activated_items.update({p: (True if i in dict_['evaluated_pipes_indices'] else False)
+        activated_items.update({p: i in dict_['evaluated_pipes_indices']
                                 for i, p in enumerate(workflow.pipes)})
 
         var_indices = []
@@ -1470,7 +1465,7 @@ class WorkflowState(DessiaObject):
                 var_indices.append(tuple(variable_indices))  # json serialisation loses tuples
             else:
                 var_indices.append(variable_indices)
-        activated_items.update({v: (True if workflow.variable_indices(v) in var_indices else False)
+        activated_items.update({v: workflow.variable_indices(v) in var_indices
                                 for v in workflow.variables})
 
         return cls(workflow=workflow, input_values=input_values, activated_items=activated_items,
@@ -1686,7 +1681,7 @@ class WorkflowState(DessiaObject):
         If state is complete, returns a WorkflowRun
         """
         if self.progress == 1:
-            values = {v: self.values[v] for v in self.workflow.variables if v.memorize and v in self.values}
+            values = {v: self.values[v] for v in self.workflow.variables if v in self.values}
             return WorkflowRun(workflow=self.workflow, input_values=self.input_values, output_value=self.output_value,
                                values=values, activated_items=self.activated_items, start_time=self.start_time,
                                end_time=self.end_time, log=self.log, name=name)
@@ -1779,7 +1774,7 @@ class WorkflowRun(WorkflowState):
         dict_['object_class'] = 'dessia_common.workflow.core.WorkflowRun'
 
         # TODO REMOVING THIS TEMPORARLY TO PREVENT DISPLAYS TO BE LOST WITH POINTERS
-        # VARIABLE_VALUES ARE NOT SET BACK IN DICT_TO_OBJECT
+        #  VARIABLE_VALUES ARE NOT SET BACK IN DICT_TO_OBJECT
         # if use_pointers:
         #     variable_values = {}
         #     for key, value in self.variable_values.items():
@@ -1800,19 +1795,41 @@ class WorkflowRun(WorkflowState):
 
         # Find & order displayable blocks
         d_blocks = [b for b in self.workflow.blocks if hasattr(b, 'display_') and hasattr(b, "_display_settings")]
-        # Change last line to isinstance ?
+        # Change last line to isinstance ? Not possible because of circular imports ?
         sorted_d_blocks = sorted(d_blocks, key=lambda b: b.order)
         for block in sorted_d_blocks:
             block_index = self.workflow.blocks.index(block)
-            settings = block._display_settings(block_index)  # Code intel is not working properly here
+            local_values = {}
+            for i, input_ in enumerate(block.inputs):
+                input_adress = self.workflow.variable_indices(input_)
+                local_values[input_] = self.variable_values[input_adress]
+            settings = block._display_settings(block_index, local_values)  # Code intel is not working properly here
             if settings is not None:
-                display_settings.append(settings)
+                display_settings.extend(settings)
 
         if isinstance(self.output_value, DessiaObject):
-            output_display_settings = [ds.compose('output_value') for ds in self.output_value.display_settings()]
+            output_display_settings = [ds.compose(attribute='output_value', serialize_data=True)
+                                       for ds in self.output_value.display_settings()]
             display_settings.extend(output_display_settings)
 
         return display_settings
+
+    def _get_from_path(self, path: str):
+        """
+        Extracts subobject at given path. Tries the generic function, then applies specific cases if it fails.
+        Returns found object
+        """
+        try:
+            return DessiaObject._get_from_path(self, path)
+        except ExtractionError:
+            segments = path.split("/")
+            first_segment = segments[1]
+            if first_segment == "values" and len(segments) == 3:
+                varindex = int(segments[2])
+                variable = self.workflow.variables[varindex]
+                upstream = self.workflow.get_upstream_nbv(variable)
+                return self.values[upstream]
+        raise NotImplementedError(f"WorkflowRun : Specific object from path method is not defined for path '{path}'")
 
     def _display_from_selector(self, selector: str, **kwargs) -> DisplayObject:
         """
@@ -1821,6 +1838,10 @@ class WorkflowRun(WorkflowState):
         # TODO THIS IS A TEMPORARY DIRTY HOTFIX OVERWRITE.
         #  WE SHOULD IMPLEMENT A WAY TO GET RID OF REFERENCE PATH WITH URLS
         track = ""
+        if "reference_path" in kwargs:
+            refpath = kwargs["reference_path"]
+        else:
+            refpath = ""
         if selector in ["documentation", "workflow"]:
             return self.workflow._display_from_selector(selector)
 
@@ -1830,16 +1851,23 @@ class WorkflowRun(WorkflowState):
         # Displays for blocks (getting reference path from block_display return)
         display_setting = self._display_settings_from_selector(selector)
         try:
-            data, reference_path = attrmethod_getter(self, display_setting.method)(**display_setting.arguments)
+            if display_setting.method == "block_display":
+                # Specific hotfix : we propagate reference_path through block_display method
+                display_object, refpath = attrmethod_getter(self, display_setting.method)(**display_setting.arguments)
+                data = display_object["data"]
+            else:
+                # But not when calling result objects display methods.
+                # We end up here when evaluating output value display
+                data = attrmethod_getter(self, display_setting.method)(**display_setting.arguments)
         except:
-            data, reference_path = None, ""
+            data = None
             track = tb.format_exc()
 
         if display_setting.serialize_data:
             data = serialize(data)
-        return DisplayObject(type_=display_setting.type, data=data, reference_path=reference_path, traceback=track)
+        return DisplayObject(type_=display_setting.type, data=data, reference_path=refpath, traceback=track)
 
-    def block_display(self, block_index: int):
+    def block_display(self, block_index: int, display_index: int):
         """
         Computes the display of associated block to use integrate it in the workflow run displays
         """
@@ -1856,7 +1884,7 @@ class WorkflowRun(WorkflowState):
             if i == block._displayable_input:
                 reference_path = f'variable_values/{input_adress}'
         display_ = block.display_(local_values=local_values, reference_path=reference_path)
-        return display_, reference_path
+        return display_[display_index], reference_path
 
     def dict_to_arguments(self, dict_: JsonSerializable, method: str):
         if method in self._allowed_methods: