+++ scripts/simple_functions_workflow.py
@@ -7,8 +7,10 @@ A simple workflow composed of functions
 import math
 import dessia_common.workflow as workflow
 
+
 def sinus_f(x: float) -> float:
     return math.sin(x)
+
 
 def asin_f(x: float) -> float:
     return math.asin(x)
@@ -23,4 +25,4 @@ workflow = workflow.WorkFlow([sinus, arcsin], [pipe1])
 
 workflow.plot_graph()
 
-workflow_run = workflow.run([math.pi/3])
+workflow_run = workflow.run([math.pi / 3])