+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Prevent foreach name from being it's iter_input name
 - Temporarly remove workflow-state from workflow run display settings
 
 ## v0.8.0